rcom: rename lib to librcom, separate applications from lib code. Fix private / public include file access. Move headers from source directory.
