+++ applications/rcdiscover/test/CMakeLists.txt
@@ -26,7 +26,7 @@ add_executable(rcom_rcdiscover_unit_tests ${SRCS})
 target_link_libraries(  rcom_rcdiscover_unit_tests
                         gtest
                         gmock
-                        rcomlib
+                        librcom
                         stdc++fs)
 
 add_test(
