+++ applications/rclaunch/CMakeLists.txt
@@ -2,14 +2,15 @@ cmake_minimum_required(VERSION 3.10)
 
 include_directories(include)
 
-add_executable(rcom
-                src/rcom.c)
+add_executable(rclaunch
+                src/rclaunch.c
+                src/run.c)
 
-target_link_libraries(rcom
+target_link_libraries(rclaunch
                         librcom)
 
 INSTALL(TARGETS
-            rcom
+            rclaunch
         DESTINATION
             "bin")
