+++ include/private/streamerlink_priv.h
@@ -21,27 +21,25 @@
   <http://www.gnu.org/licenses/>.
 
  */
-#ifndef _RCOM_RESPONSE_PRIV_H_
-#define _RCOM_RESPONSE_PRIV_H_
+#ifndef _RCOM_STREAMERLINK_PRIV_H_
+#define _RCOM_STREAMERLINK_PRIV_H_
 
-#include "rcom/response.h"
-#include "http.h"
+#include "streamerlink.h"
 
-enum {
-        RESPONSE_PARSE_HEADERS,
-        RESPONSE_PARSE_ALL
-};
+#ifdef __cplusplus
+extern "C" {
+#endif
 
+streamerlink_t *new_streamerlink(streamerlink_ondata_t ondata,
+                                 streamerlink_onresponse_t onresponse,
+                                 void* data, int autoconnect);
+        
+void delete_streamerlink(streamerlink_t *link);
 
-typedef int (*response_onheaders_t)(void *userdata, response_t *response);
-typedef int (*response_ondata_t)(void *userdata, response_t *response,
-                                 const char *data, int len);
+int streamerlink_set_remote(streamerlink_t *link, addr_t *addr);
 
-int response_set_onheaders(response_t *r, response_onheaders_t onheaders, void *userdata);
-int response_set_ondata(response_t *r, response_ondata_t ondata, void *userdata);
+#ifdef __cplusplus
+}
+#endif
 
-int response_parse_html(response_t *r, tcp_socket_t socket, int what);
-list_t *response_headers(response_t *r);
-int response_send(response_t *r, tcp_socket_t client_socket);
-
-#endif // _RCOM_RESPONSE_PRIV_H_
+#endif // _RCOM_STREAMERLINK_PRIV_H_
