+++ include/private/util_priv.h
@@ -21,26 +21,28 @@
   <http://www.gnu.org/licenses/>.
 
  */
-#ifndef _RCOM_H_
-#define _RCOM_H_
+#ifndef _RCOM_UTIL_H_
+#define _RCOM_UTIL_H_
 
-#include <rcom/app.h>
-#include <rcom/addr.h>
-#include <rcom/data.h>
-#include <rcom/circular.h>
-#include <rcom/request.h>
-#include <rcom/response.h>
-#include <rcom/registry.h>
-#include <rcom/datalink.h>
-#include <rcom/datahub.h>
-#include <rcom/messagelink.h>
-#include <rcom/messagehub.h>
-#include <rcom/service.h>
-#include <rcom/client.h>
-#include <rcom/streamer.h>
-#include <rcom/streamerlink.h>
-#include <rcom/multipart_parser.h>
-#include <rcom/dump.h>
-#include <rcom/util.h>
+#include <stdint.h>
+#include <r.h>
+#include "util.h"
 
-#endif // _RCOM_H_
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+membuf_t *escape_string(const char* s);
+
+char *encode_base64(const unsigned char *s, int len);
+        
+unsigned char *SHA1(const unsigned char* s, unsigned char *digest);
+
+int urlencode(const unsigned char* s, membuf_t *buf);
+int urldecode(const char* s, int len, membuf_t *buf);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // _RCOM_UTIL_H_
