+++ src/streamer.c
@@ -23,9 +23,9 @@
  */
 #include <r.h>
 
-#include "rcom/app.h"
-#include "rcom/dump.h"
-#include "rcom/util.h"
+#include "app.h"
+#include "dump.h"
+#include "util.h"
 
 #include "http_parser.h"
 #include "http.h"
